requirements.txt
absl-py==0.6.1
astor==0.7.1
cloudpickle==0.6.1
cycler==0.10.0
dask==0.20.2
decorator==4.3.0
gast==0.2.0
grpcio==1.16.1
grpcio-tools==1.16.1
h5py==2.9.0
html5lib==0.9999999
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
Markdown==3.0.1
matplotlib==3.0.2
networkx==2.2
numpy==1.15.4
pandas==0.23.4
Pillow==5.3.0
protobuf==3.6.1
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2018.7
PyWavelets==1.0.1
PyYAML==5.1
scikit-image==0.14.1
scikit-learn==0.19.1
scipy==1.2.1
setuptools==39.1.0
six==1.12.0
tensorboard==1.9.0
tensorflow==1.9.0
termcolor==1.1.0
toolz==0.9.0
Werkzeug==0.14.1